Add voronoi polygons to stations
From all points (given through latitude and longitude) we can calculate the
Voronoi diagram. With that we get a polygon for each weather station to
determine which one is the nearest for a given point or an area. This
calculation is done through scipy see http://docs.scipy.org/doc/scipy-dev/
reference/generated/scipy.spatial.Voronoi.html.